LogViz/requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter
dj-database-url>=2.0.0

# Task Queue (optional for now, can add later)
# celery>=5.3.0
# redis>=4.5.0

# Data Processing (from original project)
networkx>=3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# LLM Integration
openai>=1.0.0  # or anthropic, langchain, etc.
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0
